refactor: extract shared message parser and fix 11 client bugs
Phase 1 — Bug fixes:
- Fix stats f64-to-i64 truncation through entire pipeline (tbuy/tsell)
- Add parse error counting + sampled logging to both adapters
- Extract CANDLE_TF_FILTER constant, add missing candle health in native
- Increase web WS buffer to 128KB with truncation detection
- Fix heatmap/VPVR max_price init (-F64_MAX), extract normalize_unix_seconds
- Add crisp-edges CSS, orderbook "Waiting..." placeholder
- Clear and re-populate demo stores on offline stream switch

Phase 2 — Structural refactoring:
- Extract shared message parser (message_parser.odin, 406 LOC) eliminating
  ~400 LOC of duplicated parsing between native and web adapters
- Unify staging structs: both adapters now use services.Parsed_* types
  with direct struct copy instead of field-by-field assignment
- Deduplicate subject builder (channel_to_stream_parts helper)
- Remove dead code: hello_widget.odin, ~270 LOC unused legacy MM types
- Consolidate magic numbers into named constants (backoff, staging caps)

Net: -740 LOC across platform adapters, both odin check targets pass.

+++ client/src/core/widgets/trade_counter.odin
@@ -35,8 +35,8 @@ trade_counter :: proc(buf: ^ui.Command_Buffer, data: Trade_Counter_Data) {
 	for s in data.stats {
 		fx := f64(s.unix)
 		if fx < data.x_min || fx > data.x_max do continue
-		limit = math.max(limit, f64(s.tbuy))
-		limit = math.max(limit, f64(s.tsell))
+		limit = math.max(limit, s.tbuy)
+		limit = math.max(limit, s.tsell)
 	}
 
 	// Clip to viewport.
@@ -65,7 +65,7 @@ trade_counter :: proc(buf: ^ui.Command_Buffer, data: Trade_Counter_Data) {
 		cx := vp.pos.x + t * vp.size.x - bar_w * 0.5
 
 		// Buy bar (green, grows upward from center).
-		buy_h := f32(f64(s.tbuy) / limit) * half_h
+		buy_h := f32(s.tbuy / limit) * half_h
 		if buy_h > 0.5 {
 			ui.push(buf, ui.Cmd_Rect_Filled{
 				rect  = {pos = {cx, center_y - buy_h}, size = {bar_w, buy_h}},
@@ -74,7 +74,7 @@ trade_counter :: proc(buf: ^ui.Command_Buffer, data: Trade_Counter_Data) {
 		}
 
 		// Sell bar (red, grows downward from center).
-		sell_h := f32(f64(s.tsell) / limit) * half_h
+		sell_h := f32(s.tsell / limit) * half_h
 		if sell_h > 0.5 {
 			ui.push(buf, ui.Cmd_Rect_Filled{
 				rect  = {pos = {cx, center_y}, size = {bar_w, sell_h}},
